Spring cleaning, using asdf and starship
Removes loads of the unused bin files, and totally removes the .bashrc,
since all I really need is in the interactve shell. This will now use
starship for the prompt (since it's at least as good as I really need)
and asdf for managing programming language installations, since that
seems to handle the majority of what I need.